BUG/MAJOR: mux_quic: Possible crashes with quiche POST requests
When receiving data already received and before calling ncb_add()
the length of the STREAM data must be shorten and the pointer data
must be moved forward.

This bug finally could be easily reproducible with such a quiche request:
    $ cargo run --bin quiche-client -- https://127.0.0.1:8080/1024.html
      --no-verify -n 10 --method POST --body /var/www/html/1024.html
and patching haproxy to split a call to qcc_recv() in two parts to
simulate the receipt of already received STREAM data.

Must be backported to 2.6.