MINOR: quic: do not interrupt recv on empty datagram
recvmsg system call may return 0 if an empty UDP datagram has been read.
This should not be considered as an error value.

Before this patch, a 0 return value would have interrupt recvmsg loop in
the same way as with negative error codes. This is now changed : receive
procedure can now continue even after reading an empty datagram.

The precise impact of this patch is unknown. This may fix stalled
transfers. An empty datagram reception scenario has not been tested.

This should be backported up to 2.6.