BUG/MEDIUM: quic: mark FD as cannot receive on EAGAIN
Use fd_cant_recv() on EAGAIN recv for the QUIC listener socket. The file
descriptor will be polled again. This prevents the io-handler to be
called too frequently and the FD goes back

This should reduce CPU consumption as quic_sock_fd_iocb() will be called
less frequently.

This bug has been introduced by the following patch
  97ecc7a8ea5339a753507c3d4e4cd83028c6d038
  MEDIUM: quic: retrieve frontend destination address

This should be backported up to 2.6.