MINOR: quic: refactor datagram redispatch
Refactor QUIC datagram dispatch to the receiver thread. The first part
of the function quic_lstnr_dgram_dispatch() has been extracted in a new
function quic_init_dgram().

The most notable improvment is that now quic_lstnr_dgram_dispatch() is
guaranteed to succeed, which simplify its invocation : no need to call
b_sub() on receiver buffer in case of error. As a counterpart, the
caller must ensure that it provides a valid datagram. This job can be
done in part through the new function quic_init_dgram().

This should be backported up to 2.7.