MEDIUM: mux-quic: drain data on transport error
Handle an error at transport layer in two stages from the MUX
perspective. When the error is detected, flag QC_CF_ERR_CONN is set and
all future sendings are interrupted.

At this stage, MUX wait to ensure all incoming data are drained. Once a
qc_recv() has not proceed any new data, the flag QC_CF_CONN_SHUT is set.
Now, the connection is considered dead and remaining streams will be
notified of the incoming error. When the last stream is closed, MUX
instance will be freed.

This should be backported up to 2.7.