DOC: Attempt to fix dconv parsing error for tune.h2.fe.initial-window-size
It appears that dconv dislikes the "see also" part being on the same line as
the regular paragraph. The beginning of the line does not show up in the
rendered version.

Attempt to fix this by inserting an additional newline which is consistent with
other options.

(cherry picked from commit bd88998678997b30fe84858d76177ce558c090c6)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>