REGTESTS: h1_host_normalization : Add a barrier to not mix up log messages
Depending on the timing, time to time, the log messages can be mixed. A
client can start and be fully handled by HAProxy (including its log message)
before the log message of the previous client was emitted or received.  To
fix the issue, a barrier was added to be sure to eval the "expect" rule on
logs before starting the next client.

(cherry picked from commit 8fe3a33328eed8e9203058b55f9d22572ca085fa)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>