DOC: ssl: Add ocsp-update troubleshooting clues and emphasize on crt-list only aspect
The current limitation of the 'ocsp-update' option and the fact that it
can only be used in crt-lists was puzzling for some people so the doc
was amended to emphasize this specificity. A configuration extract was
added as well.
A few troubleshooting clues were added as well.

Must be backported in 2.8.

(cherry picked from commit a9b1a4edd9a0acbae42ed8f28d9f48cae1b4ef99)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>