BUG/MINOR: config: fix stick table duplicate name check
When a stick-table is defined within a peers section, the name is
prefixed with the peers section name. However when checking for
duplicate table names, the check was using the table name without
the prefix, and would thus never match.

Must be backported as far as 2.6.

(cherry picked from commit bce0ca696cffcfcd0153e76415905444571381a7)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>