BUG/MINOR: quic: Wrong Retry paquet version field endianess
The 32-bits version field of the Retry paquet was inversed by the code. As this
field must be in the network byte order on the wire, this code has supposed that
the sender of the Retry packet will always be little endian. Hopefully this is
often the case on our Intel machines ;)

Must be backported as far as 2.6.

(cherry picked from commit 5997d18c7862e9461a005a4e74ca1ab80c49bb20)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>