BUG/MEDIUM: quic: token IV was not computed using a strong secret
Computing the token key and IV, a stronger derived key was used
to compute the key but the weak secret was still used to compute
the IV. This could be used to found the secret.

This patch fix this using the same derived key than the one used
to compute the token key.

This should backport until v2.6

(cherry picked from commit 075b8f4cd897e8aab682a54db4f1dcd0b1411167)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>