BUG/MINOR: log: free errmsg on error in cfg_parse_log_forward()
When leaving cfg_parse_log_forward() on error paths, errmsg which is local
to the function could still point to valid data, and it's our
responsibility to free it.

Instead of freeing it everywhere it is invoved, we free it prior to
leaving the function.

This should be backported as far as 2.4.

(cherry picked from commit 9f9d557468befde23be828df5f35f529079bebd0)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>