MINOR: cpuset: add cpu_map_configured() to know if a cpu-map was found
Since we'll soon want to adjust the "thread-groups" degree of freedom
based on the presence of cpu-map, we first need to be able to detect
if cpu-map was used. This function scans all cpu-map sets to detect if
any is present, and returns true accordingly.

(cherry picked from commit 71344176135b6e140ebc7c107ffe40ff9c3f79ba)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>